本次更新 ·设备匿名注册 ·用户绑定设备 ·一些逻辑细节优化

+++ internal/handler/factory/DeviceFactory.go
@@ -22,8 +22,8 @@ func (d deviceFactory) CreateInstance(name, deviceType, description, ownerUUID s
 		Type:         deviceType,
 		Description:  description,
 		OwnerUUID:    ownerUUID,
-		AddTime:      int(time.Now().Unix()),
-		LastSeen:     int(time.Now().Unix()),
+		AddTime:      int64(time.Now().Unix()),
+		LastSeen:     int64(time.Now().Unix()),
 		Activated:    false,
 		Properties:   model.InstanceProperties{Items: make(map[string]model.PropertyItem)},
 	}
